Doing final changes.

+++ python/tremppi/clean.py
@@ -1,6 +1,7 @@
 import os
-from tremppi.header import database_file, widgets
-
+import sqlite3
+from .header import database_file, widgets
+from .configure import configure
 
 def clean_setup(data_path, widget):
     with open(os.path.join(data_path, widget + ".js"), 'w+') as file_js:
@@ -23,12 +24,27 @@ def delete_all_files(data_path):
         if os.path.isdir(os.path.join(data_path, widget)):
             clean_data_files(data_path, widget)
 
+
 def clean(data_path, widget):
     if widget == "editor":
-        os.remove(os.path.join(data_path, database_file))
-        clean_data("properties")
-        clean_data("select")
+        if os.path.exists(os.path.join(data_path, database_file)):
+            os.remove(os.path.join(data_path, database_file))
+        clean_data(data_path, "properties")
+        clean_data(data_path, "select")
         delete_all_files(data_path)
     if widget == "properties":
         # set all the cost+robustness+witness data
         delete_all_files(data_path)
+        clean_data(data_path, "select")
+        with sqlite3.connect(os.path.join(data_path, database_file)) as conn:
+            conn.execute('DROP TABLE IF EXISTS Properties')
+
+    clean_data_files(data_path, "qualitative")
+    clean_data_files(data_path, "quantitative")
+    clean_data_files(data_path, "regulations")
+    clean_data_files(data_path, "correlations")
+    clean_data_files(data_path, "group")
+    clean_data_files(data_path, "witness")
+
+    for widget_to_conf in widgets:
+        configure(data_path, widget_to_conf)
